+++ src/main/java/org/example/massaToppings/Portuguesa.java
@@ -6,7 +6,7 @@ public class Portuguesa extends MassaDecorator {
         super(massa);
     }
 
-    public float getPrecoAdicional() {
+    public double getPrecoAdicional() {
         return 15.0f;
     }
