+++ src/main/java/org/example/tamanhoTipoMassa/TipoMassa.java
@@ -4,5 +4,5 @@ public interface TipoMassa {
 
     String getTipoMassa();
 
-    float modificadorPreco();
+    double modificadorPreco();
 }
